+++ app/cookbooks/ssh/resources/ssh.rb
@@ -10,7 +10,8 @@ property :name,                     String, name_property: true
 property :user,                     String, default: node[:ssh][:init][:user]
 property :group,                    String, default: node[:ssh][:init][:user]
 property :vagrant_key,              String, default: node[:ssh][:vagrant_key]
-property :configuration,            Hash
+property :private_keys_list,        Array,  default: node[:ssh][:private_keys][:files]
+property :authorized_keys_list,     Array,  default: node[:ssh][:authorized_keys][:files]
 
 action :stop_ssh_agent do
     execute "Stop ssh-agent and previously-active keys" do
@@ -21,6 +22,7 @@ end
 action :clear_ssh_directory do
     execute "Clear /home/#{new_resource.user}/.ssh directory" do
         command "rm -rf /home/#{new_resource.user}/.ssh/*"
+        only_if { ::Dir.exist?("/home/#{new_resource.user}/.ssh") }
     end
 end
 
@@ -31,34 +33,35 @@ action :create_ssh_config do
         owner "#{new_resource.user}"
         group "#{new_resource.group}"
         mode "600"
+        only_if { ::Dir.exist?("/home/#{new_resource.user}/.ssh") }
     end
 end
 
 action :add_public_keys do
-    unless new_resource.configuration[:authorized_keys].nil?
+    unless new_resource.authorized_keys_list.nil?
         template "Authorized keys configuration" do
             source "authorized_keys.erb"
             path "/home/#{new_resource.user}/.ssh/authorized_keys"
-            owner "#{new_resource.user}"
-            group "#{new_resource.group}"
+            owner new_resource.user
+            group new_resource.group
             mode "600"
             sensitive true
             variables ({ 
-                vagrant_key: "#{new_resource.vagrant_key}",
-                authorized_keys: new_resource.configuration[:authorized_keys] 
+                vagrant_key: new_resource.vagrant_key,
+                authorized_keys_list: new_resource.authorized_keys_list 
             })
         end
     end
 end
 
 action :add_private_keys do
-    unless new_resource.configuration[:private_keys].nil?
-        new_resource.configuration[:private_keys].each do |private_key|
+    unless new_resource.private_keys_list.nil?
+        new_resource.private_keys_list.each do |private_key|
             cookbook_file "Adding private key file : #{private_key}" do
                 source "keys/private/#{private_key}"
                 path "/home/#{new_resource.user}/.ssh/#{private_key}"
-                owner "#{new_resource.user}"
-                group "#{new_resource.group}"
+                owner new_resource.user
+                group new_resource.group
                 mode "0400"
                 sensitive true
                 action :create_if_missing
